Move inspect_message to generators folder
