+++ kybra/compiler/kybra_generate/src/generators/canister_methods/mod.rs
@@ -1,2 +1,3 @@
 pub mod heartbeat;
 pub mod init;
+pub mod inspect_message;
